added dummy data to seed file to populate database before we actually use object to populate.

+++ db/migrate/20190326165026_create_artists_table.rb
@@ -2,7 +2,7 @@ class CreateArtistsTable < ActiveRecord::Migration[5.0]
   def change
     create_table :artists do |table|
       table.string :name
-      table.integer :popularity
+      table.float :popularity
       table.string :venue
     end
   end
